requirements_core.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
python-json-logger>=2.0.7
colorama>=0.4.6
tqdm>=4.66.1
typing-extensions>=4.10.0
aiohttp>=3.9.1
aiosignal>=1.3.1
async-timeout>=4.0.3
attrs>=23.2.0

# AI and ML
torch>=2.0.0
transformers>=4.38.1
sentence-transformers>=3.4.1
huggingface-hub>=0.19.0
accelerate>=0.20.3
safetensors>=0.4.0

# Google AI/Gemini dependencies
google-ai-generativelanguage>=0.4.0
google-cloud-aiplatform>=1.38.1
google-generativeai>=0.3.2
vertexai>=0.0.1

# Vector Databases
faiss-cpu>=1.9.0
qdrant-client>=1.7.0
pymilvus>=2.4.2
chromadb>=0.4.22

# Language Framework
langchain>=0.3.14
langchain-community>=0.3.14
langchain-core>=0.3.30
langchain-text-splitters>=0.3.5
langchain-anthropic>=0.0.1
langsmith>=0.2.11

# API and Model Access
openai>=1.12.0
anthropic>=0.8.1

# Web Scraping and Crawling
beautifulsoup4>=4.12.3
scrapy>=2.11.0
httpx>=0.28.1
httpx-sse>=0.4.0

# Data Processing
numpy>=1.26.4
pandas>=2.2.3
scikit-learn>=1.6.1
scipy>=1.15.1
spacy>=3.7.2
nltk>=3.9.1
vaderSentiment>=3.3.2

# Web/API
fastapi>=0.115.0
uvicorn>=0.25.0
websockets>=12.0.0
pydantic-settings>=2.1.0

# Miscellaneous utilities
PyYAML>=6.0.1
pillow>=10.2.0
pytz>=2024.1
matplotlib>=3.9.2
seaborn>=0.13.1
joblib>=1.4.2

# Utilities
tqdm>=4.67.1
joblib>=1.4.2
aiofiles>=23.2.1

# Web Server
fastapi>=0.115.6
uvicorn>=0.34.0
httptools>=0.6.4
watchfiles>=1.0.4

# Monitoring and Logging
prometheus-client>=0.21.1
sentry-sdk>=1.5.0

# Audio Processing (Voice Features)
torch>=2.0.0
torchaudio>=2.2.1
soundfile>=0.12.1
sounddevice>=0.4.6
PyAudio>=0.2.13
librosa>=0.10.1

# Voice and Speech Processing
openai-whisper>=20240930
ffmpeg-python>=0.2.0

# System Acceleration
accelerate>=1.3.0
tokenizers>=0.21.0
bitsandbytes>=0.41.1

# Database
SQLAlchemy>=2.0.37

# Development
memory-profiler>=0.61.0
build>=1.2.2.post1

# Additional utilities
PyYAML>=6.0.0
coloredlogs>=15.0.1
colorlog>=6.9.0
rich>=13.0.0
